+++ codegen/src/codegen/data/field_type.rs
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use super::super::{Enum, Message, MessageOrEnum, Syntax};
+use super::{Enum, Message, MessageOrEnum, Syntax};
 use crate::{ErrorKind, Result};
 use ::puroro_protobuf_compiled::google::protobuf::field_descriptor_proto;
 use ::std::rc::{Rc, Weak};
