+++ codegen/src/codegen/gen/enum.rs
@@ -18,7 +18,7 @@
 //!  - [c++ generated code](https://developers.google.com/protocol-buffers/docs/reference/cpp-generated#enum)
 
 use super::super::util::*;
-use super::super::{Enum, PackageOrMessageExt, Syntax};
+use super::{Enum, PackageOrMessageExt, Syntax};
 use crate::syn;
 use crate::syn::{parse2, Item, ItemEnum, Path, Type};
 use crate::{ErrorKind, Result};
