+++ codegen/src/codegen/gen/oneof_field.rs
@@ -13,9 +13,9 @@
 // limitations under the License.
 
 use super::super::util::*;
-use super::super::{FieldType, LengthDelimitedType, MessageExt, OneofExt, OneofField};
 use super::field::gen_default_fn;
 use super::PackageOrMessageExt;
+use super::{FieldType, LengthDelimitedType, MessageExt, OneofExt, OneofField};
 use crate::syn::{
     parse2, Expr, ExprMethodCall, Field, Ident, ImplItemMethod, Lifetime, NamedField, PathSegment,
     Type,
